implement feature registered users

+++ src/core/tools/ItemsCleaner.php
@@ -15,10 +15,16 @@ class ItemsCleaner {
         // DELETE FROM items WHERE create_time <= '2018-11-01 10:47:00';
         $now->sub(new DateInterval('P' . $this->config->getMaxAge() . 'D'));
         $timestamp = $now->format('Y-m-d H:i:s');
-        $sql = 'DELETE FROM items WHERE create_time <= \'' . $timestamp . '\';';
+        $sql = 'DELETE FROM items WHERE owner = \'' . CreateRequestDecoderImpl::DEFAULT_OWNER . '\' AND create_time < \'' . $timestamp . '\';';
 
         if (!$this->storage->execSql($sql)) {
-            error_log('Clean-Up expired items failed: ' . $sql);
+            error_log('Clean-Up anonymous expired items failed: ' . $sql);
+        }
+
+        $sql = 'DELETE FROM items WHERE owner != \'' . CreateRequestDecoderImpl::DEFAULT_OWNER . '\' AND access_time < \'' . $timestamp . '\';';
+
+        if (!$this->storage->execSql($sql)) {
+            error_log('Clean-Up registered-user expired items failed: ' . $sql);
         }
     }
 }
